Issue 145: Organize back-end app in layers Add "application" and "infrastructure" layers
