+++ back/config/packages/acceptance/services.yaml
@@ -4,5 +4,5 @@ services:
         autoconfigure: true
         public: public
 
-    App\Repository\InMemory\UserRepository:
-    App\Domain\Repository\UserRepositoryInterface: '@App\Repository\InMemory\UserRepository'
+    App\Infrastructure\Repository\InMemory\UserRepository:
+    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Repository\InMemory\UserRepository'
