+++ back/config/packages/dev/services.yaml
@@ -4,5 +4,5 @@ services:
         autoconfigure: true
         public: false
 
-    App\Repository\Doctrine\UserRepository: ~
-    App\Domain\Repository\UserRepositoryInterface: '@App\Repository\Doctrine\UserRepository'
+    App\Infrastructure\Repository\Doctrine\UserRepository: ~
+    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Repository\Doctrine\UserRepository'
