+++ back/config/routes.yaml
@@ -1,3 +1,3 @@
 #index:
 #    path: /
-#    controller: App\Controller\DefaultController::index
+#    controller: App\Infrastructure\Controller\DefaultController::index
